test server task ownership once (GOL-387)
